Fixed Seach and refactored NavigationBar
