--- src/common/NavigationBar/Logo/index.js
@@ -1,13 +1,0 @@
-import { Wrapper, Title, CameraIcon, StyledHomeButton } from "./styled";
-
-export const NavigationLogo = ({toMovies}) => {
-
-    return (
-        <StyledHomeButton to={toMovies}>
-            <Wrapper>
-           <CameraIcon />
-            <Title>Movies Browser</Title>
-        </Wrapper>
-        </StyledHomeButton>
-    );
-};
