+++ src/common/NavigationBar/index.js
@@ -1,17 +1,8 @@
-import { Search } from "./Search";
+import LogoAndButtons from "./LogoAndButtons";
+import Search from "./Search";
 import {
   NavigationBarContainer,
   NavigationBarWrapper,
-  NavigationButtonsWrapper,
-  HeaderContainer,
-  HeaderLogo,
-  Heading,
-  Navigation,
-  NavigationItem,
-  LogoNavigationLink,
-  StyledNavigationLink,
-  SearchContainer,
-  SearchIcon,
 } from "./styled";
 
 const NavigationBar = () => {
@@ -19,40 +10,10 @@ const NavigationBar = () => {
   return (
     <NavigationBarContainer>
       <NavigationBarWrapper>
-        <NavigationButtonsWrapper>
-          <HeaderContainer>
-            <LogoNavigationLink
-              to="/movies"
-            >
-              <HeaderLogo />
-              <Heading>
-                Movies Browser
-              </Heading>
-            </LogoNavigationLink>
-          </HeaderContainer>
-          <Navigation>
-            <NavigationItem>
-              <StyledNavigationLink
-                to="/movies"
-              >
-                Movies
-              </StyledNavigationLink>
-            </NavigationItem>
-            <NavigationItem>
-              <StyledNavigationLink
-                to="/people"
-              >
-                People
-              </StyledNavigationLink>
-            </NavigationItem>
-          </Navigation>
-        </NavigationButtonsWrapper>
-        <SearchContainer>
-          <SearchIcon />
-          <Search />
-        </SearchContainer>
+        <LogoAndButtons />
+        <Search />
       </NavigationBarWrapper>
-    </NavigationBarContainer>
+    </NavigationBarContainer >
   )
 };
